DataForZ/requirements.txt
firebase-admin
google-adk
google-cloud-firestore
langgraph
langchain
langchain-google-genai
pydantic